Seemingly fixed bug. Uncooked/Cooked Patty works.
I changed the De Spawner to wait with IEnumerator and Coroutine WaitForSeconds instead of the provided Destroy, lifetime float. I tested the same way as before and was unable to make the glitch appear. Should try to make the cooked patties stay on the socket(uncooked layer mask) until grabbed from grill.